+++ runtime_audit.go
@@ -651,10 +651,6 @@ func (a *runtimeAudit) recordStackEquivHashMismatch() {
 	}
 }
 
-func (a *runtimeAudit) recordStackEquivStateMismatch() {
-	a.recordStackEquivStateMismatchAt(-1)
-}
-
 func (a *runtimeAudit) recordStackEquivStateMismatchAt(depthFromTop int) {
 	if a == nil || !a.equivEnabled {
 		return
@@ -665,10 +661,6 @@ func (a *runtimeAudit) recordStackEquivStateMismatchAt(depthFromTop int) {
 		state.stackEquivStateMismatch++
 		recordStackEquivMismatchDepth(&state.stackEquivStateMismatchDepthSum, &state.stackEquivStateMismatchMaxDepth, &state.stackEquivStateMismatchDepthBuckets, depthFromTop)
 	}
-}
-
-func (a *runtimeAudit) recordStackEquivPayloadMismatch() {
-	a.recordStackEquivPayloadMismatchAt(-1)
 }
 
 func (a *runtimeAudit) recordStackEquivPayloadMismatchAt(depthFromTop int) {